+++ .editorconfig
@@ -1,0 +1,14 @@
+root = true
+
+[*]
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+charset = utf-8
+
+[{*.js,*.ts,package.json,*.yml,*.cjson}]
+indent_style = space
+indent_size = 4
+
+[*.md]
+trim_trailing_whitespace = false⏎
